requirements.txt
python-telegram-bot
sqlalchemy
aiosqlite
python-dotenv
cryptography